Refactoring: prepare for reorganisation of FileMgr and SynthHelper
these two are also rather low hanging fuit.
Both are not real objects, insofar they don't hold any state.
Rather, these are collections of utility functions, and can thus
be converted to namespaces with inline functions rather mechanically
