+++ src/LV2_Plugin/CMakeLists.txt
@@ -45,7 +45,7 @@ file (GLOB yoshimi_interface_files
     ../Interface/MidiDecode.cpp ../Interface/MidiDecode.h
     ../Interface/InterChange.cpp ../Interface/InterChange.h
     ../Interface/Data2Text.cpp ../Interface/Data2text.h
-    ../Interface/RingBuffer.cpp ../Interface/FileMgr.cpp
+    ../Interface/RingBuffer.cpp
     ../Interface/MidiLearn.cpp ../Interface/MidiLearn.h
     ../UI/MiscGui.cpp ../UI/MiscGui.h)
 file (GLOB yoshimi_params_files
